Moving mocking back into `generate_artparse.py`
